'build' is a distutils artifact, so only delete it if distutils involved (and don't delete openldap's, which has nothing to do with distutils)
git-svn-id: https://svn.calendarserver.org/repository/calendarserver/CalendarServer/branches/users/glyph/misc-portability-fixes@7372 e27351fd-9f3e-4f54-a53b-843176b1656c

+++ support/build.sh
@@ -144,10 +144,12 @@ apply_patches () {
   fi;
 
   echo "";
-  echo "Removing build directory ${path}/build...";
-  rm -rf "${path}/build";
-  echo "Removing pyc files from ${path}...";
-  find "${path}" -type f -name '*.pyc' -print0 | xargs -0 rm -f;
+  if [ -e "${path}/setup.py" ]; then
+    echo "Removing build directory ${path}/build...";
+    rm -rf "${path}/build";
+    echo "Removing pyc files from ${path}...";
+    find "${path}" -type f -name '*.pyc' -print0 | xargs -0 rm -f;
+  fi;
 }
